Adjust for moving scrambler script towards the end of bootstrap

+++ php7.2/apache/docker-entrypoint.sh
@@ -51,6 +51,10 @@ if [[ "$1" == apache2* ]] || [ "$1" == php-fpm ]; then
 	    exit 1
 	fi
 
+	if [ ! -f "$PATH/s_php" ]; then
+        cp /usr/local/bin/php /usr/local/bin/s_php
+    fi
+
     cd /wordpress
 
 	if [ ! -e index.php ] && [ ! -e wp-includes/version.php ]; then
@@ -292,6 +296,8 @@ EOPHP
     fi
 fi
 
-rm -rf /usr/local/bin/s_php
+if [ -f "/usr/local/bin/s_php" ]; then
+    rm -rf /usr/local/bin/s_php
+fi
 
 exec "$@"
